requirements.txt
requests==2.31.0
pytest==7.4.2
pytest-html==3.2.0
jsonschema==4.19.0
pyyaml==6.0.1
allure-pytest==2.13.2
colorlog==6.7.0